refactor(prep): #537 — extract scripts/prep_application.py to findajob.prep package
Third M3 extraction. Import-only — no behavior change, no LOC
consolidation.

Layout (per #538 — orchestrator + role_runner + docx_postprocess; the
proposed validation.py is dropped because extracting the inline
validation block from main() is internal restructuring, not import-
only):
- findajob/prep/orchestrator.py — main(), abbrev_title(), notify()
- findajob/prep/role_runner.py — run_role() (byte-equivalent dupe of
  the copy in scripts/interview_prep.py; consolidation deferred to
  M3+ cleanup PR)
- findajob/prep/docx_postprocess.py — _add_cover_letter_spacing(),
  _linkify_contact_info()

Module-load `load_env()` moved into main() so `findajob.prep.*` is
import-safe. Three other duplications stay through this PR per import-
only discipline:
- abbrev_title() also lives in scripts/rename_folders.py
- notify() also lives in findajob.triage.orchestrator and
  scripts/interview_prep.py
- run_role() also lives in scripts/interview_prep.py (will become
  findajob.interview.role_runner in PR #4)

scripts/prep_application.py shrinks 682 → 45 LOC (under ≤50 cap).
Module sizes:
  orchestrator.py 548, docx_postprocess.py 73, role_runner.py 70,
  __init__.py 4. orchestrator over the 300 soft cap; internal
  decomposition of main() out of M3 import-only scope.

Tests:
- New tests/test_prep_import_safety.py (6 tests) — locks the import-
  safety invariant via load_env spy on the orchestrator import; smoke-
  tests every module's public symbols; abbrev_title behavior preserved
  (Output Folder Format rule from CLAUDE.md); _linkify_contact_info
  behavior preserved.
- tests/test_prep_application_cost_logging.py — import path updated
  (`from findajob.prep.role_runner import run_role`); patch target
  updated (`findajob.prep.role_runner.complete`). Test logic
  unchanged.

The import-safety test deliberately reimports ONLY orchestrator (which
had the original `load_env()` at module load), not role_runner or
docx_postprocess. Reimporting role_runner invalidates the
`from findajob.prep.role_runner import run_role` references taken by
test_prep_application_cost_logging at collection — the closure binds
to the old module's globals and `patch()` on the new module misses.

CI green locally: 1860 passed, 1 skipped; ruff check + ruff format
--check both clean repo-wide; mypy clean on the new package.
