+++ src/findajob/prep/__init__.py
@@ -1,0 +1,4 @@
+"""Application-materials prep — orchestrate briefing, resume, cover letter, critique.
+
+Entry point: scripts/prep_application.py (a thin shim that imports from this package).
+"""
