+++ tests/test_prep_application_cost_logging.py
@@ -1,4 +1,8 @@
-"""Cost-logging tests for the run_role() helper in scripts/prep_application.py.
+"""Cost-logging tests for the run_role() helper.
+
+Post-#537: `run_role()` lives in `findajob.prep.role_runner` (extracted
+from `scripts/prep_application.py` in M3 PR #3). The test surface is
+unchanged — only the import path and the patch target moved.
 
 HTTP-mocked tests for the wrapper-driven run_role() helper after the Phase 2 port.
 Each call to run_role() writes a cost_log row with API-authoritative cost_usd
@@ -15,7 +19,7 @@ import urllib.error
 from io import BytesIO
 from unittest.mock import patch
 
-from scripts.prep_application import run_role
+from findajob.prep.role_runner import run_role
 
 # Fake key satisfies the OPENROUTER_API_KEY guard in openrouter.complete() without
 # a real network call — used in conjunction with the urlopen mock.
@@ -162,7 +166,7 @@ def test_run_role_passes_cached_prefix_to_wrapper():
     conn = sqlite3.connect(":memory:")
     conn.row_factory = sqlite3.Row
     conn.executescript(COST_LOG_SCHEMA)
-    with patch("scripts.prep_application.complete", _fake_complete):
+    with patch("findajob.prep.role_runner.complete", _fake_complete):
         run_role(
             "resume_tailor",
             "tailor this",
